requirements.txt
torch
transformers
opencv-python
numpy
pandas
scikit-learn
Pillow
